Removed testing for other apps and moods / added binary successes
Justin Tomlinson

Everything was broken before, but now that I don't need to consider
moods or third-party apps (for now) I removed them and things seem
to be working again. I also added a way to count that if a picture
was able to bypass Android Face Unlock even once, then it counts
as a success.