Created an app to take pictures for me
Justin Tomlinson
March 5, 2020

Let me start by saying there is absolutely no documentation or
comments on this app whatsoever, so don't get your hopes up to
understand it easily. That being said, this app will detect a
person's face and once they are at the correct distance(s), it
will take a picture and tell them to move to the correct next
position. There is also a manual image capture button in case
face recognition does not work. The face recognition is heavily
based on this:
https://github.com/husaynhakeem/android-face-detector